+++ src/Akeneo/Pim/Structure/Bundle/Query/PublicApi/Attribute/Sql/GetAllBlacklistedAttributeCodes.php
@@ -2,9 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Structure\Bundle\Query\InternalApi\Attribute;
+namespace Akeneo\Pim\Structure\Bundle\Query\PublicApi\Attribute\Sql;
 
-use Akeneo\Pim\Structure\Component\Query\InternalApi\GetAllBlacklistedAttributeCodesInterface;
+use Akeneo\Pim\Structure\Component\Query\PublicApi\Attribute\GetAllBlacklistedAttributeCodesInterface;
 use Doctrine\DBAL\Connection;
 
 final class GetAllBlacklistedAttributeCodes implements GetAllBlacklistedAttributeCodesInterface
