+++ src/Akeneo/Pim/Structure/Bundle/Resources/config/queries.yml
@@ -118,6 +118,6 @@ services:
             - '@database_connection'
 
     akeneo.pim.structure.query.get_all_blacklisted_attribute_codes:
-        class: Akeneo\Pim\Structure\Bundle\Query\InternalApi\Attribute\GetAllBlacklistedAttributeCodes
+        class: Akeneo\Pim\Structure\Bundle\Query\PublicApi\Attribute\Sql\GetAllBlacklistedAttributeCodes
         arguments:
             - '@database_connection'
